Create CONTRIBUTING.md
It does not appear that there is a CONTRIBUTING.md file for this project at the present time. Unfortunately, this means that new users who are looking to add to this project may not do so because they do not know where to begin. I hope that this CONTRIBUTING.md file can serve as a base for this project to build upon.