Update one method name
I think this method's name is a little bit not to make sense. All other method name is related to write, but this one is writer. It makes the name become a little bit confusing. I just change it to afterWrite which matches the previous ones that make more sense.